Network and Core modules created. Files moved from one place to another
