+++ network/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id 'com.android.application'
+    id 'com.android.library'
     id 'kotlin-android'
     id 'kotlin-kapt'
     id 'dagger.hilt.android.plugin'
@@ -10,43 +10,39 @@ android {
     buildToolsVersion "29.0.3"
 
     defaultConfig {
-        applicationId "com.gsrg.luasdublin"
         minSdkVersion 21
         targetSdkVersion 29
         versionCode 1
         versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles "consumer-rules.pro"
     }
+
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
     kotlinOptions {
-        jvmTarget = JavaVersion.VERSION_1_8
-    }
-
-    packagingOptions {
-        exclude 'META-INF/DEPENDENCIES'
-    }
-
-    buildFeatures {
-        viewBinding = true
+        jvmTarget = '1.8'
     }
 
     buildTypes {
         release {
             buildConfigField "Boolean", "MOCK_RESPONSE", "false"
+            buildConfigField "String", "LUAS_BASE_URL", "\"https://luasforecasts.rpa.ie/xml/\""
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
         debug {
             buildConfigField "Boolean", "MOCK_RESPONSE", "false"
+            buildConfigField "String", "LUAS_BASE_URL", "\"https://luasforecasts.rpa.ie/xml/\""
             minifyEnabled false
             debuggable = true
         }
         mock {
             buildConfigField "Boolean", "MOCK_RESPONSE", "true"
+            buildConfigField "String", "LUAS_BASE_URL", "\"https://luasforecasts.rpa.ie/xml/\""
             minifyEnabled false
             debuggable = true
             signingConfig signingConfigs.debug
@@ -55,52 +51,33 @@ android {
 }
 
 dependencies {
-    implementation fileTree(dir: "libs", include: ["*.jar"])
-    implementation project(path: ':core')
     implementation project(path: ':domain')
-    implementation project(path: ':network')
-    implementation project(path: ':database')
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation "androidx.core:core-ktx:$core_ktx_version"
     implementation "androidx.appcompat:appcompat:$appcompat_version"
-    implementation 'com.google.android.material:material:1.2.1'
-    implementation 'androidx.constraintlayout:constraintlayout:2.0.2'
-    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
 
-    // Hilt dependencies
-    implementation "com.google.dagger:hilt-android:$hilt_version"
-    kapt "com.google.dagger:hilt-android-compiler:$hilt_version"
-
-    // Hilt ViewModel extension
-    def hilt_jetpack_version = "1.0.0-alpha02"
-    implementation "androidx.hilt:hilt-lifecycle-viewmodel:$hilt_jetpack_version"
-    kapt "androidx.hilt:hilt-compiler:$hilt_jetpack_version"
-
-    // Navigation component
-    implementation "androidx.navigation:navigation-fragment-ktx:$nav_version"
-    implementation "androidx.navigation:navigation-ui-ktx:$nav_version"
-    implementation "androidx.navigation:navigation-dynamic-features-fragment:$nav_version"
-
+    // Retrofit
+    def retrofit_version = '2.9.0'
+    implementation "com.squareup.retrofit2:retrofit:$retrofit_version"
+    implementation "com.squareup.retrofit2:adapter-rxjava2:$retrofit_version"
+    implementation "com.squareup.retrofit2:converter-gson:$retrofit_version"
     implementation('com.squareup.retrofit2:converter-simplexml:2.6.1') {
         exclude module: 'stax'
         exclude module: 'stax-api'
         exclude module: 'xpp3'
     }
+    implementation "com.squareup.okhttp3:logging-interceptor:4.9.0"
+
+    // Coroutines
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$kotlin_corountines_version"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$kotlin_corountines_version"
 
     //RXJava
     implementation "io.reactivex.rxjava2:rxjava:$rxjava_version"
-    implementation "io.reactivex.rxjava2:rxandroid:$rxandroid_version"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$kotlin_corountines_version"
 
-    // Timber - Log
-    implementation "com.jakewharton.timber:timber:$timber_version"
-
-    // Testing Navigation
-    androidTestImplementation "androidx.navigation:navigation-testing:$nav_version"
-
-    // Hilt testing dependencies
-    androidTestImplementation "com.google.dagger:hilt-android-testing:$hilt_version"
-    kaptAndroidTest "com.google.dagger:hilt-android-compiler:$hilt_version"
+    // Hilt dependencies
+    implementation "com.google.dagger:hilt-android:$hilt_version"
+    kapt "com.google.dagger:hilt-android-compiler:$hilt_version"
 
     testImplementation "junit:junit:$junit_version"
     androidTestImplementation "androidx.test.ext:junit:$androidx_junit_version"
