+++ network/src/mock/assets/MarlboroughResponse.xml
