+++ settings.gradle
@@ -1,3 +1,5 @@
+include ':core'
+include ':network'
 include ':domain'
 include ':database'
 include ':app'
